+++ Sources/Swift/Tools/SentryLogger.swift
@@ -1,3 +1,5 @@
+@_implementationOnly import _SentryPrivate
+
 import Foundation
 
 /// **EXPERIMENTAL** - A structured logging API for Sentry.
@@ -178,22 +180,44 @@ public final class SentryLogger: NSObject {
         guard let batcher else {
             return
         }
-        var logAttributes = attributes.mapValues { SentryLog.Attribute(value: $0) }
         
+        var logAttributes = attributes.mapValues { SentryLog.Attribute(value: $0) }
+        addDefaultAttributes(to: &logAttributes)
+
         if !arguments.isEmpty {
             logAttributes["sentry.message.template"] = .string(body)
         }
         for (index, argument) in arguments.enumerated() {
             logAttributes["sentry.message.parameter.\(index)"] = SentryLog.Attribute(value: argument)
         }
-        
-        let log = SentryLog(
-            timestamp: dateProvider.date(),
-            level: level,
-            body: format(body: body, arguments: arguments),
-            attributes: logAttributes
+
+        let propagationContextTraceIdString = hub.scope.propagationContextTraceIdString
+        let propagationContextTraceId = SentryId(uuidString: propagationContextTraceIdString)
+
+        batcher.add(
+            SentryLog(
+                timestamp: dateProvider.date(),
+                traceId: propagationContextTraceId,
+                level: level,
+                body: format(body: body, arguments: arguments),
+                attributes: logAttributes
+            )
         )
-        batcher.add(log)
+    }
+
+    private func addDefaultAttributes(to attributes: inout [String: SentryLog.Attribute]) {
+        guard let batcher else {
+            return
+        }
+        attributes["sentry.sdk.name"] = .string(SentryMeta.sdkName)
+        attributes["sentry.sdk.version"] = .string(SentryMeta.versionString)
+        attributes["sentry.environment"] = .string(batcher.options.environment)
+        if let releaseName = batcher.options.releaseName {
+            attributes["sentry.release"] = .string(releaseName)
+        }
+        if let span = hub.scope.span {
+            attributes["sentry.trace.parent_span_id"] = .string(span.spanId.sentrySpanIdString)
+        }
     }
     
     private func format(body: String, arguments: [CVarArg]) -> String {